requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pytest==8.4.2
paho-mqtt==2.1.0
